intel-compute-runtime: Autospec creation for version 19.50.15079
Required update to work with gmmlib-19.4.1 release.  As gmmlib is
trying to fix their ABI/API issues they came with a few modifications
that compute-runtime requires to adapt to.

Since a new release is not available and compute-runtime is broken, a
couple of patches were backported to easily address this difference

Signed-off-by: Daniel Charles <daniel.charles@intel.com>